\name{print.fourier.descriptor}
\alias{print.fourier.descriptor}
\title{
Print a Fourier descriptor object.
}
\usage{
fourier.descriptor <- random.fourier.descriptor(12)

print(fourier.descriptor)}
\description{
Print a Fourier descriptor object.
}
\arguments{
	\item{fourier.descriptor}{A Fourier descriptor object.}
}
\examples{
library('FourierDescriptors')

fourier.descriptor <- random.fourier.descriptor(12)

print(fourier.descriptor)}
